Added return reason popup form
When user clicks return button, a form is first prompted to them asking for return reason.
This info is attached to the ticket object.
The return reason and additional comment are displayed in the ticket card.
Modified StyledSelect component to have a default option.
Modified the TicketModel to include the ticket return reason and additional comments field