Adjusted get events and get event routes
Added isClaimable & userHasTicket fields to individual event objects.
This allows for displaying UI that tells user that they have a ticket or not and also changes the claim ticket button based on isClaimable field.
Event cards will display a badge (Tickets unavailable or Ticket claimed) based on event conditions associated with the user.